Implement subscription resources, accomodate to schema mode PHR-14610
Add two new resource types, AidboxSubscriptionTopic and AidboxTopicDestination
in order to support Aidbox topic-based subscriptions feature https://docs.aidbox.app/modules/topic-based-subscriptions/wip-dynamic-subscriptiontopic-with-destinations
Update the docker image to a more recent version since it was very old (and
also didn't have this feature).
Update the SearchParameter resource, this was previously a custom aidbox format
but schema - which is required for subscriptions - mode makes it use and
validate the FHIR spec.
Surface the response body on failures to terraform so the user is displayed
what's wrong, since we only print status codes and can't access the actual
underlying problem as an end-user.
api_client.go was getting a bit out of hand so it got simplified